Refactor in the favorite package
