+++ backend-boxboxd/src/main/java/com/kropsz/github/backendboxboxd/service/favorite/factory/ObjectFactory.java
@@ -1,4 +1,4 @@
-package com.kropsz.github.backendboxboxd.util.factory;
+package com.kropsz.github.backendboxboxd.service.favorite.factory;
 
 public interface ObjectFactory {
     Object create (String id, Long userId);
